+++ git-proc/src/add.rs
@@ -43,8 +43,8 @@ impl<'a> Add<'a> {
     }
 
     /// Execute the command and return the exit status.
-    pub fn status(self) -> Result<(), CommandError> {
-        crate::Build::build(self).status()
+    pub async fn status(self) -> Result<(), CommandError> {
+        crate::Build::build(self).status().await
     }
 }