+++ git-proc/src/config.rs
@@ -38,8 +38,8 @@ impl<'a> Config<'a> {
     }
 
     /// Execute the command and return the exit status.
-    pub fn status(self) -> Result<(), CommandError> {
-        crate::Build::build(self).status()
+    pub async fn status(self) -> Result<(), CommandError> {
+        crate::Build::build(self).status().await
     }
 }