+++ git-proc/src/push.rs
@@ -58,8 +58,8 @@ impl<'a> Push<'a> {
     }
 
     /// Execute the command and return the exit status.
-    pub fn status(self) -> Result<(), CommandError> {
-        crate::Build::build(self).status()
+    pub async fn status(self) -> Result<(), CommandError> {
+        crate::Build::build(self).status().await
     }
 }