+++ git-proc/src/worktree.rs
@@ -119,8 +119,8 @@ impl<'a> Add<'a> {
     }
 
     /// Execute the command and return the exit status.
-    pub fn status(self) -> Result<(), CommandError> {
-        crate::Build::build(self).status()
+    pub async fn status(self) -> Result<(), CommandError> {
+        crate::Build::build(self).status().await
     }
 }
 
@@ -181,8 +181,8 @@ impl<'a> Remove<'a> {
     }
 
     /// Execute the command and return the exit status.
-    pub fn status(self) -> Result<(), CommandError> {
-        crate::Build::build(self).status()
+    pub async fn status(self) -> Result<(), CommandError> {
+        crate::Build::build(self).status().await
     }
 }